feat(api): follow & unfollow

+++ api/src/auth/auth.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotAcceptableException } from '@nestjs/common';
+import { Injectable, NotAcceptableException, NotFoundException } from '@nestjs/common';
 import { AuthTokens, LoginDto, SignupDto } from '../types';
 import { hash, compare } from 'bcrypt';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -13,7 +13,7 @@ export class AuthService {
     this.prisma = new PrismaService();
     this.utils = new UtilsService();
   }
-  async signup(signupDto: SignupDto): Promise<AuthTokens> {
+  async signup(signupDto: SignupDto): Promise<any> {
     const transactionMeter: { transacted: boolean; userId: string | null } = {
       transacted: false,
       userId: null,
@@ -45,6 +45,18 @@ export class AuthService {
           password: await this.hash(signupDto.password),
           refreshToken: '',
         },
+        select : {
+          id : true ,
+          username : true ,
+          email : true ,    
+          profile : {
+            select : {
+              profilePicture : true ,
+              bio : true ,
+              status :  true,
+            }
+          }
+        }
       });
       transactionProxy.transacted = true;
       transactionProxy.userId = user.id;
@@ -56,7 +68,7 @@ export class AuthService {
         where: { id: user.id },
         data: { refreshToken: tokens.refreshToken },
       });
-      return tokens;
+      return { tokens : { accessToken : tokens.accessToken, refreshToken : tokens.refreshToken} , user : user};
     } catch (err: any) {
       if (err instanceof PrismaClientKnownRequestError) console.log(err);
       if (transactionProxy.transacted) {
@@ -104,6 +116,7 @@ export class AuthService {
       },
     });
     if (!user) throw new NotAcceptableException('Invalid credentials');
+    if(!await this.prisma.user.findUnique({ where : { id   }})) throw new NotFoundException("Sorry, we couldn't find that user");
     await this.prisma.user.delete({
       where: { id: id },
     });
